requirements.txt
dashscope>=1.14.0
fastmcp>=0.1.0
mcp>=1.0.0
langchain-mcp-adapters>=0.1.0
langchain-core>=0.3.0
deepagents
python-dotenv
gradio
